Remove the OverlayBlocker from users plugin pass set lockApp in the strapi object
