+++ packages/strapi-admin/admin/src/containers/App/actions.js
@@ -5,32 +5,24 @@
  */
 
 import {
+  FREEZE_APP,
   LOAD_PLUGIN,
-  UPDATE_PLUGIN,
-  PLUGIN_LOADED,
   PLUGIN_DELETED,
+  PLUGIN_LOADED,
+  UNFREEZE_APP,
   UNSET_HAS_USERS_PLUGIN,
+  UPDATE_PLUGIN,
 } from './constants';
+
+export function freezeApp() {
+  return {
+    type: FREEZE_APP,
+  };
+}
 
 export function loadPlugin(newPlugin) {
   return {
     type: LOAD_PLUGIN,
-    plugin: newPlugin,
-  };
-}
-
-export function updatePlugin(pluginId, updatedKey, updatedValue) {
-  return {
-    type: UPDATE_PLUGIN,
-    pluginId,
-    updatedKey,
-    updatedValue,
-  };
-}
-
-export function pluginLoaded(newPlugin) {
-  return {
-    type: PLUGIN_LOADED,
     plugin: newPlugin,
   };
 }
@@ -42,8 +34,30 @@ export function pluginDeleted(plugin) {
   };
 }
 
+export function pluginLoaded(newPlugin) {
+  return {
+    type: PLUGIN_LOADED,
+    plugin: newPlugin,
+  };
+}
+
+export function unfreezeApp() {
+  return {
+    type: UNFREEZE_APP,
+  };
+}
+
 export function unsetHasUserPlugin() {
   return {
     type: UNSET_HAS_USERS_PLUGIN,
+  };
+}
+
+export function updatePlugin(pluginId, updatedKey, updatedValue) {
+  return {
+    type: UPDATE_PLUGIN,
+    pluginId,
+    updatedKey,
+    updatedValue,
   };
 }
