+++ packages/strapi-admin/admin/src/containers/App/constants.js
@@ -4,8 +4,10 @@
  *
  */
 
-export const UNSET_HAS_USERS_PLUGIN = 'app/App/UNSET_HAS_USERS_PLUGIN';
+export const FREEZE_APP = 'app/App/FREEZE_APP';
 export const LOAD_PLUGIN = 'app/App/LOAD_PLUGIN';
-export const UPDATE_PLUGIN = 'app/App/UPDATE_PLUGIN';
 export const PLUGIN_LOADED = 'app/App/PLUGIN_LOADED';
 export const PLUGIN_DELETED = 'app/App/PLUGIN_DELETED';
+export const UNFREEZE_APP = 'app/App/UNFREEZE_APP';
+export const UNSET_HAS_USERS_PLUGIN = 'app/App/UNSET_HAS_USERS_PLUGIN';
+export const UPDATE_PLUGIN = 'app/App/UPDATE_PLUGIN';
