+++ packages/strapi-admin/admin/src/containers/App/reducer.js
@@ -7,16 +7,19 @@ import {
 import { fromJS } from 'immutable';
 
 import {
-  UPDATE_PLUGIN,
+  FREEZE_APP,
   PLUGIN_DELETED,
   PLUGIN_LOADED,
+  UNFREEZE_APP,
   UNSET_HAS_USERS_PLUGIN,
+  UPDATE_PLUGIN,
 } from './constants';
 
 const initialState = fromJS({
-  showGlobalAppBlocker: true,
-  plugins: {},
+  blockApp: false,
   hasUserPlugin: true,
+  plugins: {},
+  showGlobalAppBlocker: true,
 });
 
 function appReducer(state = initialState, action) {
@@ -25,12 +28,16 @@ function appReducer(state = initialState, action) {
       return state.set('showGlobalAppBlocker', false);
     case ENABLE_GLOBAL_OVERLAY_BLOCKER:
       return state.set('showGlobalAppBlocker', true);
+    case FREEZE_APP:
+      return state.set('blockApp', true);
     case PLUGIN_LOADED:
       return state.setIn(['plugins', action.plugin.id], fromJS(action.plugin));
     case UPDATE_PLUGIN:
       return state.setIn(['plugins', action.pluginId, action.updatedKey], fromJS(action.updatedValue));
     case PLUGIN_DELETED:
       return state.deleteIn(['plugins', action.plugin]);
+    case UNFREEZE_APP:
+      return state.set('blockApp', false);
     case UNSET_HAS_USERS_PLUGIN:
       return state.set('hasUserPlugin', false);
     default:
