+++ packages/strapi-helper-plugin/lib/src/actions/overlayBlocker.js
@@ -8,8 +8,6 @@
 import {
   DISABLE_GLOBAL_OVERLAY_BLOCKER,
   ENABLE_GLOBAL_OVERLAY_BLOCKER,
-  FREEZE_APP,
-  UNFREEZE_APP,
 } from '../constants/overlayBlocker';
 
 export function disableGlobalOverlayBlocker() {
@@ -21,17 +19,5 @@ export function disableGlobalOverlayBlocker() {
 export function enableGlobalOverlayBlocker() {
   return {
     type: ENABLE_GLOBAL_OVERLAY_BLOCKER,
-  };
-}
-
-export function freezeApp() {
-  return {
-    type: FREEZE_APP,
-  };
-}
-
-export function unfreezeApp() {
-  return {
-    type: UNFREEZE_APP,
   };
 }
