+++ packages/strapi-helper-plugin/lib/src/constants/overlayBlocker.js
@@ -6,5 +6,3 @@
 
 export const DISABLE_GLOBAL_OVERLAY_BLOCKER = 'strapiHelperPlugin/OverlayBlocker/DISABLE_GLOBAL_OVERLAY_BLOCKER';
 export const ENABLE_GLOBAL_OVERLAY_BLOCKER = 'strapiHelperPlugin/OverlayBlocker/ENABLE_GLOBAL_OVERLAY_BLOCKER';
-export const FREEZE_APP = 'strapiHelperPlugin/OverlayBlocker/FREEZE_APP';
-export const UNFREEZE_APP = 'strapiHelperPlugin/OverlayBlocker/UNFREEZE_APP';
