+++ packages/strapi-helper-plugin/lib/src/utils/request.js
@@ -1,6 +1,6 @@
 import 'whatwg-fetch';
 import auth from 'utils/auth';
-import { freezeApp, unfreezeApp } from 'actions/overlayBlocker';
+
 /**
  * Parses the JSON returned by a network request
  *
@@ -87,7 +87,8 @@ function serverRestartWatcher(response) {
       }
     })
       .then(() => {
-        unfreezeApp();
+        // Hide the global OverlayBlocker
+        strapi.unlockApp();
         resolve(response);
       })
       .catch(err => {
@@ -143,7 +144,8 @@ export default function request(url, options = {}, shouldWatchServerRestart = fa
     .then(parseJSON)
     .then((response) => {
       if (shouldWatchServerRestart) {
-        freezeApp();
+        // Display the global OverlayBlocker
+        strapi.lockApp();
         return serverRestartWatcher(response);
       }
