+++ packages/strapi-plugin-users-permissions/admin/src/containers/App/actions.js
@@ -3,20 +3,3 @@
  * App actions
  *
  */
-
-import {
-  FREEZE_APP,
-  UNFREEZE_APP,
-} from './constants';
-
-export function freezeApp() {
-  return {
-    type: FREEZE_APP,
-  };
-}
-
-export function unfreezeApp() {
-  return {
-    type: UNFREEZE_APP,
-  };
-}
