+++ packages/strapi-plugin-users-permissions/admin/src/containers/App/constants.js
@@ -3,6 +3,3 @@
  * App constants
  *
  */
-
-export const FREEZE_APP = 'UsersPermissions/App/FREEZE_APP';
-export const UNFREEZE_APP = 'UsersPermissions/App/UNFREEZE_APP';
