+++ packages/strapi-plugin-users-permissions/admin/src/containers/App/reducer.js
@@ -5,21 +5,11 @@
  */
 
 import { fromJS } from 'immutable';
-import {
-  FREEZE_APP,
-  UNFREEZE_APP,
-} from './constants';
 
-const initialState = fromJS({
-  blockApp: false,
-});
+const initialState = fromJS({});
 
 function appReducer(state = initialState, action) {
   switch (action.type) {
-    case FREEZE_APP:
-      return state.set('blockApp', true);
-    case UNFREEZE_APP:
-      return state.set('blockApp', false);
     default:
       return state;
   }
