+++ packages/strapi-plugin-users-permissions/admin/src/containers/App/selectors.js
@@ -1,16 +1,7 @@
-import { createSelector } from 'reselect';
+// import { createSelector } from 'reselect';
 
 /**
  * Direct selector to the list state domain
  */
 
-const selectGlobalDomain = () => state => state.get('global');
-
-const makeSelectBlockApp = () => createSelector(
-  selectGlobalDomain(),
-  (globalState) => globalState.get('blockApp'),
-);
-
-export {
-  makeSelectBlockApp,
-};
+// const selectGlobalDomain = () => state => state.get('global');
