+++ packages/strapi-plugin-users-permissions/admin/src/containers/EditPage/saga.js
@@ -9,8 +9,6 @@ import {
   takeLatest,
 } from 'redux-saga/effects';
 
-import { freezeApp, unfreezeApp } from 'containers/App/actions';
-
 import request from 'utils/request';
 
 import {
@@ -92,8 +90,6 @@ export function* roleGet(action) {
 
 export function* submit() {
   try {
-    yield put(freezeApp());
-
     const actionType = yield select(makeSelectActionType());
     const body = yield select(makeSelectModifiedData());
     const roleId = yield select(makeSelectRoleId());
@@ -106,12 +102,10 @@ export function* submit() {
     const response = yield call(request, requestURL, opts, true);
 
     if (response.ok) {
-      yield put(unfreezeApp());
       yield put(submitSucceeded());
     }
   } catch(error) {
     console.log(error);
-    yield put(unfreezeApp());
   }
 }
